requirements.txt
alembic==1.11.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-SSLify==0.1.5
PyMySQL==1.1.0
SQLAlchemy==2.0.17
Werkzeug==2.3.6
python-dotenv==1.0.0
Flask-Cors==4.0.0
requests==2.26.0 